requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2022.1
Django==4.2.2
django-crispy-forms==2.0
django-etc==1.4.0
django-hitcount==1.3.5
django-resized==1.0.2
django-taggit==4.0.0
Pillow==9.5.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
